Change interface of gl-draw-elements etc. for the consistency
The C API takes mode, count, type, and pointer to index array.  If we give
an index array directly, we can deduce count and type from it, so
we don't need them.  If we use data in VBO, however, we need them;
previously we made count and type optional to handle those two cases,
but it made the latter call awkward, for the order of arguments
doesn't match C API.

We revised it so that it works polymorphically:
 (gl-draw-elements mode index-vec)  ; direct data
 (gl-draw-elements mode count type [offset]) ; use VBO

Same change for gl-draw-range-elements and gl-draw-elements-base-vertex.

NB: gl-vertex-pointer family has a similar issue, but in their case
we also need to support passing u8vector _AND_ type.  We'll think over.